removed old files, refactored views
-Added missing entity/model commitAuthor
