+++ lib/core/injection_container.dart
@@ -1,5 +1,6 @@
 import 'package:get_it/get_it.dart';
-import 'package:git_history/services/github_service.dart';
+
+import '../features/Github/data/datasources/remote/github_service.dart';
 
 final getIt = GetIt.instance;
