+++ lib/features/Github/domain/entities/commit_author_entity.dart
@@ -1,65 +1,11 @@
-import 'package:equatable/equatable.dart';
+class CommitAuthorEntity {
+  final String? name;
+  final String? email;
+  final DateTime? date;
 
-class CommitAuthorEntity extends Equatable {
-  final String? login;
-  final int? id;
-  final String? nodeId;
-  final String? avatarUrl;
-  final String? gravatarId;
-  final String? url;
-  final String? htmlUrl;
-  final String? followersUrl;
-  final String? followingUrl;
-  final String? gistsUrl;
-  final String? starredUrl;
-  final String? subscriptionsUrl;
-  final String? organizationsUrl;
-  final String? reposUrl;
-  final String? eventsUrl;
-  final String? receivedEventsUrl;
-  final String? type;
-  final bool? siteAdmin;
-
-  const CommitAuthorEntity({
-    this.login,
-    this.id,
-    this.nodeId,
-    this.avatarUrl,
-    this.gravatarId,
-    this.url,
-    this.htmlUrl,
-    this.followersUrl,
-    this.followingUrl,
-    this.gistsUrl,
-    this.starredUrl,
-    this.subscriptionsUrl,
-    this.organizationsUrl,
-    this.reposUrl,
-    this.eventsUrl,
-    this.receivedEventsUrl,
-    this.type,
-    this.siteAdmin,
+  CommitAuthorEntity({
+    required this.name,
+    required this.email,
+    required this.date,
   });
-
-  @override
-  List<Object?> get props => [
-        login,
-        id,
-        nodeId,
-        avatarUrl,
-        gravatarId,
-        url,
-        htmlUrl,
-        followersUrl,
-        followingUrl,
-        gistsUrl,
-        starredUrl,
-        subscriptionsUrl,
-        organizationsUrl,
-        reposUrl,
-        eventsUrl,
-        receivedEventsUrl,
-        type,
-        siteAdmin,
-      ];
 }
