+++ lib/features/Github/domain/repositories/commit_repository.dart
@@ -1,7 +1,8 @@
 import 'package:dartz/dartz.dart';
+
 import '../../../../core/failures/failures.dart';
+import '../../data/models/branch_model.dart';
 import '../entities/commit_history_entity.dart';
-import '../../../../models/branch_model.dart';
 
 abstract class CommitRepository {
   Future<Either<Failure, List<BranchModel>>> getBranchtList();
