+++ lib/features/Github/presentation/widgets/all_commits_failure.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:git_history/blocs/github_bloc/github_bloc.dart';
+import '../bloc/github_bloc.dart';
+
 
 class AllCommitsFailure extends StatelessWidget {
   const AllCommitsFailure({super.key});
