+++ lib/main.dart
@@ -3,8 +3,8 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:intl/date_symbol_data_local.dart';
 
-import 'blocs/github_bloc/github_bloc.dart';
 import 'core/observer/app_bloc_observer.dart';
+import 'features/Github/presentation/bloc/github_bloc.dart';
 import 'screens/landing/landing_page.dart';
 
 Future<void> main() async {
