+++ pubspec.yaml
@@ -9,6 +9,7 @@ environment:
 dependencies:
   bloc: ^8.1.2
   cupertino_icons: ^1.0.2
+  dartz: ^0.10.1
   equatable: ^2.0.5
   flutter:
     sdk: flutter
